refactor: Chooser.tsx の fetch ロジックを useResolveAbsolutePath フックに移動
resolvePathViaSidecar と resolveAbsolutePath をコンポーネントから切り離し、
useWorkspaceResources.ts の useResolveAbsolutePath フックに集約。
